+++ domain/zone/record/structs.go
@@ -2,19 +2,19 @@ package record
 
 
 type RecordInfo struct {
-    Id int
+    Id int64
     Name string
-    Ttl int
+    Ttl int64
     Type string
     Value string
 }
 
 type RecordAdd struct {
-    Id int `goptions:"-i, --id, obligatory, description='Zone id'"`
-    Version int `goptions:"-v, --version, obligatory, description='Zone version'"`
+    Id int64 `goptions:"-i, --id, obligatory, description='Zone id'"`
+    Version int64 `goptions:"-v, --version, obligatory, description='Zone version'"`
     Name string `goptions:"-n, --name, obligatory, description='Record name. Relative name, may contain leading wildcard. @ for empty name'"`
     Type string `goptions:"-t, --type, obligatory, description='Record type'"`
     Value string `goptions:"-V, --value, obligatory, description='Value for record. Semantics depends on the record type.'"`
-    Ttl int `goptions:"-T, --ttl, description='Time to live, in seconds, between 5 minutes and 30 days'"`
+    Ttl int64 `goptions:"-T, --ttl, description='Time to live, in seconds, between 5 minutes and 30 days'"`
 }
